Make slight tweaks for better mobile experience
The top corners were still rounded on the mobile view, which looked
strange. So, I modified the media query to remove the border radius
on the top left and right corners. I also lowered the max-width for the
media query - it was kicking in a few pixels after the flex-wrap, creating
a bit of a stutter in the transition to a portrait orientation.